+++ memberaudit/models/general.py
@@ -41,6 +41,7 @@ class General(models.Model):
             ("finder_access", "Can access character finder feature"),
             ("reports_access", "Can access reports feature"),
             ("characters_access", "Can view characters owned by others"),
+            ("exports_access", "Can access data exports"),
             ("view_shared_characters", "Can view shared characters"),
             ("view_same_corporation", "Can view corporation characters"),
             ("view_same_alliance", "Can view alliance characters"),
@@ -67,20 +68,12 @@ class General(models.Model):
             user.has_perm("memberaudit.view_same_alliance")
             and user.profile.main_character.alliance_id
         ):
-            return (
-                cls.users_with_basic_access()
-                .select_related("profile__main_character")
-                .filter(
-                    profile__main_character__alliance_id=user.profile.main_character.alliance_id
-                )
+            return cls.users_with_basic_access().filter(
+                profile__main_character__alliance_id=user.profile.main_character.alliance_id
             )
         elif user.has_perm("memberaudit.view_same_corporation"):
-            return (
-                cls.users_with_basic_access()
-                .select_related("profile__main_character")
-                .filter(
-                    profile__main_character__corporation_id=user.profile.main_character.corporation_id
-                )
+            return cls.users_with_basic_access().filter(
+                profile__main_character__corporation_id=user.profile.main_character.corporation_id
             )
         return User.objects.filter(pk=user.pk)