+++ memberaudit/tests/test_tasks.py
@@ -14,7 +14,9 @@ from app_utils.testing import generate_invalid_pk
 
 from ..models import Character, CharacterAsset, CharacterUpdateStatus, Location
 from ..tasks import (
+    _export_data_for_topic,
     delete_character,
+    export_data,
     run_regular_updates,
     update_all_characters,
     update_character,
@@ -666,6 +668,10 @@ class TestUpdateCharacter(TestCase):
         self.assertTrue(self.character_1001.is_update_status_ok())
 
 
+@patch(
+    TASKS_PATH + ".Character.objects.get_cached",
+    lambda pk, timeout: Character.objects.get(pk=pk),
+)
 @patch(TASKS_PATH + ".fetch_esi_status", lambda: EsiStatus(True, 99, 60))
 @patch(TASKS_PATH + ".MEMBERAUDIT_LOG_UPDATE_STATS", False)
 @patch(MODELS_PATH + ".character.MEMBERAUDIT_DATA_RETENTION_LIMIT", None)
@@ -786,3 +792,34 @@ class TestDeleteCharacter(TestCase):
         delete_character.delay(character_1001.pk)
         # then
         self.assertFalse(Character.objects.filter(pk=character_1001.pk).exists())
+
+
+@override_settings(CELERY_ALWAYS_EAGER=True, CELERY_EAGER_PROPAGATES_EXCEPTIONS=True)
+class TestExportData(TestCase):
+    @classmethod
+    def setUpClass(cls) -> None:
+        super().setUpClass()
+        load_entities()
+        cls.character = create_memberaudit_character(1001)
+
+    @patch(TASKS_PATH + ".data_exporters.export_topic_to_archive")
+    def test_should_export_all_topics(self, mock_export_topic_to_file):
+        # when
+        export_data()
+        # then
+        called_topics = [
+            call[1]["topic"] for call in mock_export_topic_to_file.call_args_list
+        ]
+        self.assertEqual(len(called_topics), 3)
+        self.assertSetEqual(
+            set(called_topics), {"contract", "contract-item", "wallet-journal"}
+        )
+
+    @patch(TASKS_PATH + ".data_exporters.export_topic_to_archive")
+    def test_should_export_wallet_journal(self, mock_export_topic_to_file):
+        # when
+        _export_data_for_topic(topic="abc")
+        # then
+        self.assertTrue(mock_export_topic_to_file.called)
+        _, kwargs = mock_export_topic_to_file.call_args
+        self.assertEqual(kwargs["topic"], "abc")