+++ memberaudit/urls.py
@@ -160,4 +160,15 @@ urlpatterns = [
         views.character_attribute_data,
         name="character_attribute_data",
     ),
+    path("data-export/", views.data_export, name="data_export"),
+    path(
+        "data-export/run-update/<str:topic>",
+        views.data_export_run_update,
+        name="data_export_run_update",
+    ),
+    path(
+        "data-export/download/<str:topic>",
+        views.download_export_file,
+        name="download_export_file",
+    ),
 ]