+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [1.9.0] - 2022-01-24
+
+### Added
+
+- Ability to download data export files directly from the web site.
+
+### Changed
+
+- Restricted access with `view_same_corporation` and `view_same_alliance` to affiliations from main character only (#106)
+- Removed support for outdated Python 3.6 & Django 3.1
+
 ## [1.8.0] - 2022-01-19
 
 ### Added